third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/shadow/suppress-mutation-events-in-shadow-characterdata-expected.txt

CONSOLE WARNING: Listener added for a 'DOMCharacterDataModified' mutation event. This event type is deprecated, and will be removed from this browser very soon. Usage of this event listener will cause performance issues today, and represents a large risk of future site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Ensures that DOMCharacterDataModified isn't fired inside shadows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fired is false
PASS div.innerHTML is 'Hello, World!'
PASS successfullyParsed is true

TEST COMPLETE
